Filter drivers by experience and Google LDAP

+++ src/com/droidsandfriends/ManageDriversServlet.java
@@ -1,11 +1,10 @@
 package com.droidsandfriends;
 
-import java.io.IOException;
-
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
 
 @SuppressWarnings("serial")
 public class ManageDriversServlet extends HttpServlet {
@@ -25,8 +24,14 @@ public class ManageDriversServlet extends HttpServlet {
 
     boolean isAscending = "1".equals(request.getParameter("a"));
     request.setAttribute("isAscending", isAscending);
-    
-    request.setAttribute("drivers", Driver.findAll(orderBy, isAscending));
+
+    String experience = request.getParameter("experience");
+    request.setAttribute("experience", experience);
+
+    boolean onlyGooglers = "on".equals(request.getParameter("onlyGooglers"));
+    request.setAttribute("onlyGooglers", onlyGooglers);
+
+    request.setAttribute("drivers", Driver.findAll(orderBy, isAscending, experience, onlyGooglers));
    
     // Render form.
     request.getRequestDispatcher("/admin/managedrivers.jsp").forward(request, response);
@@ -35,7 +40,9 @@ public class ManageDriversServlet extends HttpServlet {
   @Override
   protected void doPost(HttpServletRequest request, HttpServletResponse response)
       throws ServletException, IOException {
-    Driver.deleteByIds(request.getParameterValues("delete"));
+    if ("delete".equals(request.getParameter("action"))) {
+      Driver.deleteByIds(request.getParameterValues("delete"));
+    }
     doGet(request, response);
   }
